x/asset/types/message_send_to_algorand_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/realiotech/network/testutil/sample"
	"github.com/stretchr/testify/require"
)

func TestMsgSendToAlgorand_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgSendToAlgorand
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgSendToAlgorand{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgSendToAlgorand{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
